Decoding bytes -> strings, changing subprocess usage
Tried to get rid of all the os.system calls because using subprocess is supposedly
more Pythonic. Should double check that we are using .call when we should and
.check_output when we should. Also, need to be careful about using shell=True, depending
on where this code is being run this can be used to execute malicious code.
A solution would be to sanitize inputs to make sure they are actually only IPs